Backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
